+++ src/lib.rs
@@ -64,6 +64,7 @@ mod ucred;
 pub use listener::UnixSeqpacketListener;
 pub use socket::UnixSeqpacket;
 pub use split::{ReadHalf, WriteHalf};
+pub use ucred::UCred;
 
 /// Get the socket type for a close-on-exec non-blocking seqpacket socket.
 fn socket_type() -> socket2::Type {